TMS: Migrate DB from sqlite to postgres (#39)
* Initial postgresql DB setup script.

* Initial work on sqlite to psql migration.

* Update to sqlite to psql migration script.

* WIP initial code changes to switch to postgres. cargo build OK.

* WIP switch to posgres. New sql init file. Update code. Update scripts to reset db and local test install

* WIP switch to posgres. New sql init file. Update code. local test install runs with 1 warning

* WIP switch to posgres.

* WIP switch to posgres.

* WIP switch to posgres. local test install runs with no warnings

* WIP switch to posgres. Replace remaining ?'s in SQL. Server appears to start ok

* WIP switch to posgres. Work on main startup init and adding DB cfg options.

* WIP switch to posgres. Work on main startup init and adding DB cfg options.

* WIP switch to posgres. Get db url from env. Initial version of audit tables.

* WIP switch to posgres. Require db url from env. Work on 1002_audit migrations. -install SUCCESS

* WIP switch to posgres. Move 1002_audit migrations to bak for possible future inclusion. -install SUCCESS

* WIP switch to postgres. Refactor startup, some cleanup. -install SUCCESS

* WIP switch to postgres. Update SQL. Install and startup OK. createkey,getpubkey OK

* WIP switch to postgres. Work on script to migrate from sqlite to psql. Export with post-process looks good.

* WIP switch to postgres. Initial version of backup script to be run where server is deployed.

* WIP switch to postgres. Work on migrate scripts.

* WIP switch to postgres. Work on migrate scripts.

* WIP switch to postgres. Work on upgrade/migrate scripts.

* WIP switch to postgres. Work on upgrade/migrate scripts.

* WIP switch to postgres. Upgrade/migrate appears okay, still need to test tms_server.

* 0.3.0 Migrate from sqlite to postgres. Update test passes.

+++ deployment/native/upgrade_020_030.sh
@@ -4,38 +4,21 @@
 # This script builds a release version and updates files in the install directory as needed.
 # Default install directory is /opt/tms_server. May be overridden using env variable TMS_INSTALL_DIR.
 # User used to build and install TMS may be given on the command line. Default user is "tms"
+#
+# Assumptions:
+#  - We are running from a checkout of tms_server github repo.
+#  - Following are installed: rust tool chain (cargo, rustc), SQLite and postgres psql.
+#  - Following env variables are set:
+#    - TMS_DB_HOST     e.g. localhost
+#    - TMS_DB_PORT     e.g. 5431
+#    - TMS_DB_USER     e.g. tms
+#    - TMS_DB_PASSWORD
+#    - POSTGRES_PASSWORD
+#
+# A test mode is supported allowing for execution as a non-root user and tms_install_user is taken to be current user.
 
 PrgName=$(basename "$0")
-
-USAGE="Usage: $PrgName [ <tms_install_user> ]"
-
-# Check number of arguments
-if [ $# -gt 1 ]; then
-  echo "$USAGE"
-  exit 1
-fi
-
-# Make sure we are running as root
-if [ "$EUID" != 0 ]; then
-  echo "This program must be run as the root user"
-  echo "Exiting ..."
-  exit 1
-fi
-
-# Determine TMS install user
-INSTALL_USR=tms
-if [ -n "$1" ]; then
-  INSTALL_USR="$1"
-fi
-
-# Make sure the specified user for the TMS install exists
-id "$INSTALL_USR" #>/dev/null 2>&1
-RET_CODE=$?
-if [ $RET_CODE -ne 0 ]; then
-    echo "TMS install user does not exist. User: $INSTALL_USR"
-    echo "Exiting ..."
-    exit $RET_CODE
-fi
+USAGE="Usage: $PrgName [ <tms_install_user> | --test ]"
 
 # Determine absolute path to location from which we are running and change to that directory.
 RUN_DIR=$(pwd)
@@ -43,7 +26,49 @@ PRG_RELPATH=$(dirname "$0")
 cd "$PRG_RELPATH"/. || exit
 PRG_PATH=$(pwd)
 
-# Set installation directory
+# Upgrade script is located under deployment/native. Some operations are relative to the top level source directory.
+SRC_DIR=$PRG_PATH/../..
+
+# Check number of arguments
+if [ $# -gt 1 ]; then
+  echo "$USAGE"
+  exit 1
+fi
+
+TEST_MODE=false
+if [ "$1" == "--test" ]; then
+  echo "*******************************"
+  echo "     Running in test mode"
+  echo "*******************************"
+  TEST_MODE=true
+fi
+
+# Make sure we are running as root or are in test mode
+if [ "$TEST_MODE" == "false" ] && [ "$EUID" != 0 ]; then
+  echo "This program must be run as the root user"
+  echo "Exiting ..."
+  exit 1
+fi
+
+# Determine TMS install user
+if [ "$TEST_MODE" == "true" ]; then
+  INSTALL_USR=$USER
+elif [ -n "$1" ]; then
+  INSTALL_USR="$1"
+else
+  INSTALL_USR=tms
+fi
+
+# Make sure the specified user for the TMS install exists
+id "$INSTALL_USR" >/dev/null 2>&1
+RET_CODE=$?
+if [ $RET_CODE -ne 0 ]; then
+    echo "TMS install user does not exist. User: $INSTALL_USR"
+    echo "Exiting ..."
+    exit $RET_CODE
+fi
+
+# Set installation directory.
 INSTALL_DEF_DIR="/opt/tms_server"
 if [ -n "$TMS_INSTALL_DIR" ]; then
   INSTALL_DIR="$TMS_INSTALL_DIR"
@@ -70,10 +95,14 @@ fi
 VERS_OLD=$(cat $VERS_FILE)
 
 # Determine new version
-VERS_NEW=$(su - $INSTALL_USR -c 'cd tms_server; cargo pkgid | cut -d "#" -f2')
+if [ "$TEST_MODE" == "true" ]; then
+  VERS_NEW=$(cd $SRC_DIR; cargo pkgid | cut -d "#" -f2)
+else
+  VERS_NEW=$(su - $INSTALL_USR -c 'cd tms_server; cargo pkgid | cut -d "#" -f2')
+fi
 
 # Set path to built executable
-EXEC_FILE=$PRG_PATH/target/release/tms_server
+EXEC_FILE=$SRC_DIR/target/release/tms_server
 
 # Construct script to be used by install user to build new executable
 echo
@@ -83,7 +112,7 @@ TMP_FILE=$(mktemp)
 # Construct first part of script
 echo "#!/bin/bash" > $TMP_FILE
 # Place various env variables into script
-echo "PRG_PATH=$PRG_PATH" >> $TMP_FILE
+echo "SRC_DIR=$SRC_DIR" >> $TMP_FILE
 echo "INSTALL_DIR=$INSTALL_DIR" >> $TMP_FILE
 echo "VER_OLD=$VERS_OLD" >> $TMP_FILE
 echo "VER_NEW=$VERS_NEW" >> $TMP_FILE
@@ -94,8 +123,8 @@ echo "Upgrading TMS Server from version $VERS_OLD to version $VERS_NEW"
 echo "Install directory: $INSTALL_DIR"
 
 # Build executable
-echo "Building executable from directory: $PRG_PATH"
-cd $PRG_PATH
+echo "Building executable from directory: $SRC_DIR"
+cd $SRC_DIR
 cargo build --release
 EOB
 
@@ -107,8 +136,13 @@ echo "===== Running build script as TMS install user. User: $INSTALL_USR"
 echo "========================================================================================="
 chmod +x $TMP_FILE
 chown $INSTALL_USR:$INSTALL_USR $TMP_FILE
-su - $INSTALL_USR -c "$TMP_FILE"
-RET_CODE=$?
+if [ "$TEST_MODE" != "true" ]; then
+  su - $INSTALL_USR -c "$TMP_FILE"
+  RET_CODE=$?
+else
+  $TMP_FILE
+  RET_CODE=$?
+fi
 echo "========================================================================================="
 if [ $RET_CODE -ne 0 ]; then
   echo
@@ -125,20 +159,45 @@ if [ ! -f "$EXEC_FILE" ]; then
   exit 1
 fi
 
-# Shut down the service, copy the new executable into place, start the service
-echo
-echo "===== Stopping TMS service and copying new executable into place"
-echo "========================================================================================="
-systemctl stop tms_server
+# Shut down the service, copy the new executable into place
+if [ "$TEST_MODE" != "true" ]; then
+  echo
+  echo "===== Stopping TMS service and copying new executable into place"
+  echo "========================================================================================="
+  systemctl stop tms_server
+fi
 cp $EXEC_FILE $INSTALL_DIR/tms_server
+
+# Update migrations files.
+BAK_TIMESTAMP=`date  +%Y%m%d%H%M%S`
+mv $TMS_HOME/migrations $TMS_HOME/migrations_bak_$BAK_TIMESTAMP
+cp -pr $SRC_DIR/resources/migrations $TMS_HOME/migrations
+chmod 0700 $TMS_HOME/migrations
+
+# Before updating version and starting up new tms_server perform the migration from sqlite to postgres
+echo
+echo "===== Migrating DB from sqlite to postgres"
+echo "========================================================================================="
+
+$SRC_DIR/migrate_to_psql/migrate_from_sqlite.sh
+RET_CODE=$?
+if [ $RET_CODE -ne 0 ]; then
+  echo
+  echo "*************** Error running migration script"
+  echo "Exiting ..."
+  exit $RET_CODE
+fi
 
 # Update version in install dir
 echo "$VERS_NEW" > $VERS_FILE
 
+# Start the service
 echo
 echo "===== Starting TMS service"
 echo "========================================================================================="
-systemctl start tms_server
+if [ "$TEST_MODE" != "true" ]; then
+  systemctl start tms_server
+fi
 
 # Remove the temporary file
 rm -f $TMP_FILE
